svm-conformance: move txn harness
