+++ dev-bins/Cargo.toml
@@ -144,6 +144,7 @@ solana-time-utils = "3.0.0"
 solana-tpu-client = { path = "../tpu-client", version = "=4.3.0-alpha.2", default-features = false, features = ["agave-unstable-api"] }
 solana-transaction = "4.1.0"
 solana-transaction-context = { path = "../transaction-context", version = "=4.3.0-alpha.2", features = ["agave-unstable-api", "bincode"] }
+solana-transaction-error = "3.3.1"
 solana-transaction-status = { path = "../transaction-status", version = "=4.3.0-alpha.2", features = ["agave-unstable-api"] }
 solana-unified-scheduler-pool = { path = "../unified-scheduler-pool", version = "=4.3.0-alpha.2", features = ["agave-unstable-api"] }
 solana-version = { path = "../version", version = "=4.3.0-alpha.2", features = ["agave-unstable-api"] }
