+++ svm-conformance/Cargo.toml
@@ -40,28 +40,34 @@ cli = ["dep:clap"]
 dev-context-only-utils = []
 dummy-for-ci-check = []
 ffi = [
+    "dep:agave-feature-set",
     "dep:agave-precompiles",
     "dep:prost",
     "dep:protosol",
     "dep:solana-compute-budget",
+    "dep:solana-instruction",
     "dep:solana-program-runtime",
     "dep:solana-pubkey",
     "dep:solana-svm",
     "dep:solana-syscalls",
+    "dep:solana-transaction-error",
     "solana-svm/conformance",
 ]
 frozen-abi = []
 
 [dependencies]
+agave-feature-set = { workspace = true, optional = true }
 agave-precompiles = { workspace = true, optional = true }
 clap = { version = "4.5.59", features = ["derive"], optional = true }
 prost = { version = "0.11", optional = true }
 protosol = { version = "=11.0.0", optional = true }
 solana-compute-budget = { workspace = true, optional = true }
+solana-instruction = { workspace = true, optional = true }
 solana-program-runtime = { workspace = true, optional = true }
 solana-pubkey = { workspace = true, optional = true }
 solana-svm = { workspace = true, optional = true }
 solana-syscalls = { workspace = true, optional = true }
+solana-transaction-error = { workspace = true, optional = true }
 
 [dev-dependencies]
 bincode = { workspace = true }
